requirements.txt
selenium
beautifulsoup4
python-telegram-bot
supabase
undetected-chromedriver